feat: organize files in mirrored src and tests dirs
